+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 Version 1.2.0-dev
 - allowed registering multiple listeners at the same time in PriorityListenerProvider
+- added default priority for listeners in PriorityListenerProvider
 
 Version 1.1.0
 - allowed registering multiple listeners at the same time in ListenerProvider