+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-Version 1.2.0-dev
+Version 1.2.0
 - allowed registering multiple listeners at the same time in PriorityListenerProvider
 - added default priority for listeners in PriorityListenerProvider
 - added DebugEventDispatcher